requirements.txt
flask
gunicorn
python-dotenv

requests
trafilatura
beautifulsoup4

pypdf
python-docx

langchain
langchain-groq
langgraph
